dtl-third-sem/backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.111.0
uvicorn==0.24.0

# Database
pymongo==4.6.3

# Data validation and serialization
pydantic==2.5.2

# Authentication and security
passlib==1.7.4
python-jose==3.3.0

# Environment variables
python-dotenv==1.0.1